+++ androidclient/app/src/main/java/pt/ipl/isel/leic/ps/androidclient/data/api/mapper/output/OutputCustomMealMapper.kt
@@ -14,7 +14,7 @@ class OutputCustomMealMapper(
 
     fun mapToOutputModel(restaurant: CustomMeal) = CustomMealOutput(
         name = restaurant.name,
-        quantity = restaurant.amount.toInt(),
+        quantity = restaurant.amount,
         unit = restaurant.unit.toString(),
         ingredients = ingredientMapper.mapToOutputModelCollection(restaurant.mealComponents.plus(restaurant.ingredientComponents)),
         cuisines = cuisineMapper.mapToOutputModelCollection(restaurant.cuisines),
