+++ androidclient/app/src/main/java/pt/ipl/isel/leic/ps/androidclient/data/api/mapper/output/OutputIngredientMapper.kt
@@ -8,7 +8,7 @@ class OutputIngredientMapper {
 
     fun mapToOutputModel(model: MealIngredient) = IngredientOutput(
         identifier = requireNotNull(model.submissionId),
-        quantity = model.amount.toInt()
+        quantity = model.amount
     )
 
     fun mapToOutputModelCollection(models: Collection<MealIngredient>) =
